added a bunch of new features and added to docs
1. remove link from last crumb automatically

This allows you to add crumbs anywhere and never have to worry
about which one is last (and which to omit the link on)

If you forget a link it get printed anyway.

2. Different styles

The crumbs can be printed as an

 <ol> (default)
 <ul>
 <p> (with a separator)

the class for the list and the separator for the <p>
can be overridden

3. Docs improved to show simpler way to add the crumbs

A before_filter in the controller adds the controller crumb
for every action making the action code DRYer.